Extracted the CSV listener logic into a separate class
