+++ test/scrutiny_runner.py
@@ -6,4 +6,3 @@
 #
 #   Copyright (c) 2021 Scrutiny Debugger
 
-
